+++ django_spire/seeding/model/enums.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class ModelSeederDefaultsEnum(str, Enum):
+    LLM = "llm"
+    FAKER = "faker"
+    INCLUDED = "included"
+